REFACTOR: routes for reactapp into its own Blueprint
